refact: 불필요한 inner class를 별도의 클래스로 분리
