+++ app/src/main/java/com/kiwi/kiwitalk/ui/search/ChatInfoViewHolder.kt
@@ -1,0 +1,20 @@
+package com.kiwi.kiwitalk.ui.search
+
+import androidx.recyclerview.widget.RecyclerView
+import com.kiwi.domain.model.ChatInfo
+import com.kiwi.domain.model.keyword.Keyword
+import com.kiwi.kiwitalk.databinding.ItemChatListNoImageBinding
+import com.kiwi.kiwitalk.ui.keyword.recyclerview.SelectedKeywordAdapter
+
+class ChatInfoViewHolder(
+    val binding: ItemChatListNoImageBinding
+) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(chatInfo: ChatInfo) {
+        binding.item = chatInfo
+        val adapter = SelectedKeywordAdapter()
+        binding.adapter = adapter
+        adapter.submitList(chatInfo.keywords.map {
+            Keyword(it, 0)
+        }.toMutableList())
+    }
+}
